fix: Falta algumas alterações para funcionar 100%
