+++ interfaces/icons__rc.py
@@ -7451,17 +7451,17 @@ qt_resource_struct = b"\
 \x00\x00\x00\x00\x00\x02\x00\x00\x00\x06\x00\x00\x00\x03\
 \x00\x00\x00\x00\x00\x00\x00\x00\
 \x00\x00\x00x\x00\x00\x00\x00\x00\x01\x00\x01\xa2:\
-\x00\x00\x01~'\xd1\xa9\xd8\
+\x00\x00\x01\x88\xc6-.\xa7\
 \x00\x00\x00Z\x00\x00\x00\x00\x00\x01\x00\x01\x9a>\
-\x00\x00\x01~'\xd1\xa9\xd8\
+\x00\x00\x01\x88\xc6-.\xab\
 \x00\x00\x00\x10\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01\x88\x9e $\xdf\
+\x00\x00\x01\x88\xc6-.\xa9\
 \x00\x00\x00D\x00\x00\x00\x00\x00\x01\x00\x01\x94\xf5\
-\x00\x00\x01~'\xd1\xa9\xd8\
+\x00\x00\x01\x88\xc6-.\xa5\
 \x00\x00\x00,\x00\x00\x00\x00\x00\x01\x00\x00,\xf5\
-\x00\x00\x01\x88\xc5mC9\
+\x00\x00\x01\x88\xc6-R\xd6\
 \x00\x00\x00\x90\x00\x00\x00\x00\x00\x01\x00\x01\xa9\xbb\
-\x00\x00\x01\x88\xc5\xb6\xa3k\
+\x00\x00\x01\x88\xc6-R\xd8\
 "
 
 def qInitResources():
